make a prompt for automatic unsecure driver uninstall

+++ Project-Aurora/Project-Aurora/Modules/HardwareMonitorModule.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel;
 using System.Threading.Tasks;
+using System.Windows;
 using AuroraRgb.Modules.HardwareMonitor;
 using AuroraRgb.Nodes;
 using AuroraRgb.Utils;
@@ -9,6 +10,8 @@ namespace AuroraRgb.Modules;
 
 public sealed class HardwareMonitorModule : AuroraModule
 {
+    private const string PromptMessage = "Would you like Aurora to automatically uninstall unsecure drivers?\n(inputx64 and WinRing0)";
+    
     protected override Task Initialize()
     {
         Global.Configuration.PropertyChanged += ConfigurationOnPropertyChanged;
@@ -17,14 +20,28 @@ public sealed class HardwareMonitorModule : AuroraModule
             LocalPcInformation.HardwareMonitor = new HardwareMonitor.HardwareMonitor();
         }
 
-        if (!Global.Configuration.EnableAmdCpuMonitor)
+        if (Global.Configuration.AutoRemoveUnsecureDrivers is null)
         {
-            UnsecureDrivers.DeleteDriver(UnsecureDrivers.InpOutDriverName, true);
+            var result = MessageBox.Show(PromptMessage, "AuroraRgb", MessageBoxButton.YesNo);
+            Global.Configuration.AutoInstallGsi = result switch
+            {
+                MessageBoxResult.Yes => true,
+                MessageBoxResult.No => false,
+                _ => true,
+            };
         }
 
-        if (!Global.Configuration.EnableWinRing0Monitor)
+        if(Global.Configuration.AutoRemoveUnsecureDrivers ?? true)
         {
-            UnsecureDrivers.DeleteDriver(UnsecureDrivers.WinRing0DriverName, true);
+            if (!Global.Configuration.EnableAmdCpuMonitor)
+            {
+                UnsecureDrivers.DeleteDriver(UnsecureDrivers.InpOutDriverName, true);
+            }
+
+            if (!Global.Configuration.EnableWinRing0Monitor)
+            {
+                UnsecureDrivers.DeleteDriver(UnsecureDrivers.WinRing0DriverName, true);
+            }
         }
 
         return Task.CompletedTask;
